+++ src/app/map/map.component.spec.ts
@@ -1,10 +1,9 @@
 import { HttpClientTestingModule } from '@angular/common/http/testing';
-import { HttpClient } from '@angular/common/http';
+import { ReactiveFormsModule} from '@angular/forms';
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { ToastrModule } from 'ngx-toastr';
 import { MapComponent } from './map.component';
 import { Workflow } from '../shared/interfaces/workflow/workflow';
-import { ConfigService } from '../shared/config/config.service';
 
 describe('MapComponent', () => {
   let component: MapComponent;
@@ -53,7 +52,7 @@ describe('MapComponent', () => {
                   ]
               },
               {
-                "text": "Query by Fire Perimeters",
+                "text": "Query by Fire Perimeter",
                 "selected": false,
                 "nestedSteps": 
                 [
@@ -88,6 +87,7 @@ describe('MapComponent', () => {
     await TestBed.configureTestingModule({
       imports: [
         HttpClientTestingModule,
+        ReactiveFormsModule,
         ToastrModule.forRoot(),
       ],
       declarations: [ MapComponent ],